requirements.txt
fastapi
pydantic
uvicorn
openai
pdfplumber
pytesseract
Pillow
faiss-cpu
streamlit
requests
openpyxl
numpy
